yolo5/requirements.txt
flask
pyyaml
loguru

# testing

pylint
pytest
boto3
pymongo
subprocess